+++ api/web/src/components/Admin/Videos/AdminVideoService.vue
@@ -1,0 +1,68 @@
+<template>
+    <div class='card-header'>
+        <div class='card-title'>Video Service</div>
+        <div class='ms-auto btn-list'>
+            <TablerIconButton
+                title='Edit Service'
+                @click='disabled = false'
+            >
+                <IconPencil :size='32' stroke='1'/>
+            </TablerIconButton>
+            <TablerIconButton
+                title='Refresh'
+                @click='fetchService'
+            >
+                <IconRefresh :size='32' stroke='1'/>
+            </TablerIconButton>
+        </div>
+    </div>
+    <div class='card-body'>
+        <TablerLoading v-if='loading' />
+        <TablerNone
+            v-else-if='!service.configured'
+            label='Video ECS Service'
+            :create='false'
+        />
+        <VideoConfig
+            v-else
+            :service='service'
+            :disabled='disabled'
+            @cancel='disabled = true'
+        />
+    </div>
+</template>
+
+<script setup lang='ts'>
+import { ref, onMounted } from 'vue';
+import { std, stdurl } from '../../../../src/std.ts';
+import VideoConfig from './VideoConfig.vue';
+import {
+    IconPencil,
+    IconRefresh
+} from '@tabler/icons-vue';
+import {
+    TablerNone,
+    TablerIconButton,
+    TablerLoading
+} from '@tak-ps/vue-tabler';
+
+const disabled = ref(true);
+const loading = ref(true);
+
+const service = ref({
+    configured: false
+});
+
+onMounted(async () => {
+    loading.value = true;
+    await fetchService()
+    loading.value = false;
+});
+
+async function fetchService() {
+    loading.value = true;
+    const url = stdurl('/api/video/service');
+    service.value = await std(url);
+    loading.value = false;
+}
+</script>
